Python log for simple heat diffusion
This is the script for meshing a cropped version of the Ti64 image. The materials are assigned to the 'black' region and the 'white' region. The material parameters (Young's modulus, Poisson ratio, thermal conductivity and expansion) are not accurate, they are just random for demonstrative purposes. The skeleton/mesh has 0.88 homogeneity index which can probably be improved in further iterations. An issue in this simulation is the smoothness of grains, if somebody could smoothen/segment the grains, I feel the simulation would look a lot better. The simulation includes heat diffusing from the left to the right of the model.