Created Python log for single grain
This file contains the simulation input for a single grain. The material parameters assigned are random, I have just tried to vary the thermal conductivities of each material to see the impact on the diffusion of heat across the grain (from left to right). The homogeneity index is ~0.99 which is a lot higher than the simple_diffusion one as the image used is a lot cleaner. I will probably try several grains in the next simulation if everything here is approved.