Uploaded Test Case 1
This is a simple binary image. It is a black star with a white background. Simple heat diffusion occurs in the simulation from the LHS to the RHS. Note that the units in the .doc file can be found in the ti6-4_material_properties.txt file under FEM. This is just to test the simulation runs successfully, colormaps will be added in further tests as will reasonable material parameters.