Uploaded Test Case 2
This is simple diffusion across a grain. The material parameters are close to that of Ti64 (units for the .doc file can be found in ti6-4_material_properties.txt under FEM). No colourmap/colourbar has been added to these images. This will be added in the next simulation. The file 'displacement.pdf' also shows the displacement in the x direction, colour bar will be added soon. Not much displacement/conductivity occurs as the thermal conductivities and thermal expansion are very similar in each phase, will change this up in next test