Added files from Test Case 2
These files include the original image/grain that was used in this simulation, the displaced mesh after the simulation, the original mesh (i.e. before the simulation), the mesh file that can be opened in Abaqus, the temperature distribution as well as the displacements in the x and y direction. The contour plots have been correctly labelled