New Simulation Results Added
The simulation was rerun for grain2. This includes a new plot for the temperature and displacement (x and y) plots. A new stress plot is also added to this test case. The original and displaced mesh is also uploaded here. Finally, as the simulation was rerun, it is likely that the mesh is not exactly the same as the previous simulation, hence why a new mesh file is uploaded, which can be viewed in Abaqus