Added Final Simulation Files
This file includes the final simulation results. It includes a temperature distribution (0-1000 degrees C) from left to right, two displacement contour maps (x and y directions), a stress trace map, the mesh file (which can be opened using Abaqus), and an image of the mesh before and after the simulation.